feat: WeeklyGiftFields component added
